Moved ingestion scripts to self contained folder. Updated documentation for Sentinel2 ingestion
